requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
streamlit 
requests
celery 
redis 
django-celery-beat
django-background-tasks